Fixed: https://github.com/sabledb-io/sabledb/issues/22
It is possible that the same composite key is overriden with a new composite key of the same type (e.g. SET) but with a different UID
An example for this can happen in the following command sequence

```
SADD set1 4 5 6
SADD set2 1 2 4 5
SADD set3 set3_item_1 set3_item_2

# store into set3 the intersection of set1 / set2, override set3 if it exists
SINTERSTORE set3 set1 set2
```

Running the above, creates a **new** `set3` of the same type (i.e. SET) with a different UID

This commit makes sure that the now orphan items {set3_item_1, set3_item_2} are purged by the cron thread