Make sure that "flushdb" is replicated
It seems that RocksDB does not create WAL entries (that can be iterated)
for `delete_range`. We workaround this by inserting special records into
the database with a special type that indicate that delete_range should
be executed on the receiving end. These entries are not visible to the
end user.

Signed-off-by: SableDB <sabledb.database@gmail.com>